改进错误处理方式；将models包里的k_project合并到projects里
