+++ models/projects.go
@@ -1,5 +1,18 @@
 package models
 
+import(
+	"errors"
+)
+
+// 导入项目时的信息
+type ProjectDetail struct {
+	ProjectName string     `gorm:"column:project_name" json:"projectName"`
+	ProjectCoverUrl string `gorm:"column:project_cover_url" json:"projectCoverUrl"`
+	Introduction string    `gorm:"column:project_description" json:"introdection"`
+	GithubUrl string       `gorm:"column:project_url" json:"githubUrl"`
+	DecideType int         `json:"decideType"` 
+}
+
 //获取每个项目的memberlist
 func GetMemberList(projects []*Project) {
 	//对列表中的每个项目查询参与该项目的全部成员
@@ -41,4 +54,19 @@ func GetManageProjects(userID string) []*Project {
 	GetMemberList(managedProjects)
 
 	return managedProjects
+}
+
+func AddProject(project *ProjectDetail) (int, error) {
+	tem := -1
+	//查询数据库是否已有该url的项目，若有则将其id保存在tem里
+	DB.Table("k_project").Select("project_id").Where("project_url = ?", project.GithubUrl).First(&tem)
+	if tem >= 0 {
+		// 说明有ID，即项目已在平台内
+		return 0, errors.New("项目已在平台内")
+	} else if err := DB.Table("k_project").Create(&project).Error; err != nil {
+		// 数据库操作错误
+		return -1, err
+	} else {
+		return 1, nil
+	}
 }
